pyproject.toml
[project]
name = "rmq-pubsub"
version = "0.1.0"
description = "Simple rabbitmq pub/sub system with 1 publisher and 2 consumers"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cassandra-driver>=3.29.2",
    "pika>=1.3.2",
]
